Reorganizing infrastructure scripts and code
