+++ gradle/rat.gradle
@@ -69,6 +69,7 @@ rat {
             '**/*.rej',
             '**/*.orig',
             '**/*.MF',
+            '**/*.cfg',
 
     ]
 }
