+++ infrastructure/scripts/aws/image/files/defaults.cfg
