+++ infrastructure/scripts/aws/image/packer.json
