+++ infrastructure/scripts/aws/launch_cluster.sh
@@ -17,8 +17,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-TAG=${1}
+set -e
 
-pushd ../../
-./gradlew destroyCluster --args "${TAG}"
+TAG=${1}
+COUNT=${2}
+
+pushd ../../../
+./gradlew launchCluster --args "${TAG} ${COUNT}"
 popd
