+++ infrastructure/scripts/aws/run_tests.sh
