+++ infrastructure/scripts/google_cloud/build_image.sh
