+++ infrastructure/scripts/google_cloud/launch_cluster.sh
